Update “Agenda” section
The term “order of the day” has a specific meaning in Robert’s Rules. In §41 of the 11th edition:

An order of the day is an item of business that is scheduled to be taken up during a given session, day, or meeting, or at a given hour (unless there is business having precedence over it that interferes).

and later:

An order of the day … is a particular subject, question, or item of business that is set in advance to be taken up during a given session, day, or meeting, or at a given hour, provided that no business having precedence over it interferes.

Also, according to Code 3‑8(b), it’s the High Alpha, not a Parliamentarian, who decides points of order.