Delete section on “Appointive” officers
Items 1, 2, and 4 of the list of appointive officers seem to be based on the 41st (2012) edition of the Statutory Code at the latest:

Code III-8. Chapter Editor and Correspondent. The High Alpha shall appoint a member to be editor of the Chapter publication and correspondent to the Cross & Crescent. It shall be his duty to edit the Chapter publication and to furnish for publication in the Cross & Crescent such material as shall be requested.

Code III-28. Duties of the Student Advisory Committee Correspondent. The High Alpha shall appoint a Student Advisory Committee Correspondent whose duty shall be to maintain liaison between his Chapter and the Student Advisory Committee member in his area.

Code III-29. Duties of the Historian. The Historian shall be appointed by the High Alpha and shall be responsible not only for maintaining the history of the Chapter, but also for ensuring that all members are educated enough about the business of the Fraternity to ensure the survival of the Chapter. The Historian may form a Committee on Chapter History, members of which would be appointed according to the Chapter’s Bylaws. The Historian and, if formed, the Committee on Chapter History shall also work in cooperation with all officers in the implementation of the Standards for Chapter Excellence.

These officers don’t appear in recent editions of the Statutory Code.

Code 3‑8(b) gives the High Alpha “the power to appoint any Chapter officers or committees not provided for by the Laws of the Fraternity or the bylaws of his Chapter,” which would cover items 3 and 5.

That leaves item 6, which sounds similar to what’s already required by Section 3.3(a)(10) of the Constitution.