Remove section describing Executive Committee membership
This appears to be closely based on Code 3‑30 of the 41st (2012) edition of the Statutory Code at the latest. The 42nd (2014) and later editions of the Statutory Code don’t include the High Theta and High Sigma as ex-officio members. No readily available edition of the Statutory Code provides for electing new ex-officio members to replace statutorily required ex-officio members. The Executive Committee is currently defined in Code 3‑9(a).