Remove “Educational Advisor” section
This appears to be based on Code 3‑15 of the 41st (2012) edition of the Statutory Code at the latest:

Each Chapter shall be encouraged to appoint an Educational Adviser [sic] who is not a Collegiate Brother or Associate Member of such Chapter, and whose duties shall be to (1) supervise the cultural, intellectual, and educational programs of the Chapter, (2) counsel the High Sigma and Educational Committee in the performance of their duties, (3) furnish such reports as may be requested of him by the Grand High Zeta, the Office of Administration, or the General Educational Adviser. He shall have further powers and duties as may be delegated by the Grand High Zeta. The Educational Adviser shall be a member of the faculty whenever possible.

All references to an “Educational Adviser” appear to have been removed in the 42nd (2014) edition of the Statutory Code.