Clarify voting on associate membership
The term “formal bid meeting” appears to be undefined; it doesn’t appear elsewhere in the bylaws or in the Constitution and Statutory Code. Also, Code 4‑3 says, “A vote on the association of a man to the Fraternity may be taken at any meeting of the Chapter, provided a quorum of both the Collegiate Brothers and Associate Members is present.” This would include chapter meetings.